fix jinja error when using url_for() in recipes
