+++ bocadillo/routing.py
@@ -1,21 +1,28 @@
 import inspect
 from functools import partial
-from typing import Callable, Union, Type, Any
-from typing import Optional, TypeVar, Generic, Dict
+from typing import (
+    Any,
+    Callable,
+    Dict,
+    Generic,
+    Optional,
+    Tuple,
+    Type,
+    TypeVar,
+    Union,
+)
 
 from parse import parse
 from starlette.websockets import WebSocketClose
 
 from . import views
-from .app_types import HTTPApp
-from .app_types import Scope, Receive, Send
+from .app_types import HTTPApp, Receive, Scope, Send
 from .errors import HTTPError
 from .redirection import Redirection
 from .request import Request
 from .response import Response
-from .views import View, HandlerDoesNotExist
-from .websockets import WebSocketView, WebSocket
-
+from .views import HandlerDoesNotExist, View
+from .websockets import WebSocket, WebSocketView
 
 # Base classes
 
@@ -30,7 +37,7 @@ class BaseRoute:
     def __init__(self, pattern: str):
         if not pattern.startswith("/"):
             pattern = f"/{pattern}"
-        self._pattern = pattern
+        self.pattern = pattern
 
     def url(self, **kwargs) -> str:
         """Return full path for the given route parameters.
@@ -43,7 +50,7 @@ class BaseRoute:
             A full URL path obtained by formatting the route pattern with
             the provided route parameters.
         """
-        return self._pattern.format(**kwargs)
+        return self.pattern.format(**kwargs)
 
     def parse(self, path: str) -> Optional[dict]:
         """Parse an URL path against the route's URL pattern.
@@ -53,7 +60,7 @@ class BaseRoute:
             If the URL path matches the URL pattern, this is a dictionary
             containing the route parameters, otherwise None.
         """
-        result = parse(self._pattern, path)
+        result = parse(self.pattern, path)
         if result is not None:
             return result.named
         return None
@@ -89,9 +96,15 @@ class BaseRouter(Generic[_R]):
     def __init__(self):
         self.routes: Dict[str, _R] = {}
 
+    def _get_key(self, route: BaseRoute) -> str:
+        raise NotImplementedError
+
     def add_route(self, *args, **kwargs):
         """Register a route. Not implemented."""
         raise NotImplementedError
+
+    def add(self, route: _R):
+        self.routes[self._get_key(route)] = route
 
     def route(self, *args, **kwargs):
         """Register a route by decorating a view."""
@@ -114,6 +127,11 @@ class BaseRouter(Generic[_R]):
                 return RouteMatch(route=route, params=params)
         return None
 
+    def mount(self, other: "BaseRouter[_R]", root: str = ""):
+        for route in other.routes.values():
+            route.pattern = root + route.pattern
+            self.add(route)
+
 
 # HTTP
 
@@ -131,12 +149,12 @@ class HTTPRoute(BaseRoute):
 
     def __init__(self, pattern: str, view: View, name: str):
         super().__init__(pattern)
-        self._view = view
-        self._name = name
+        self.view = view
+        self.name = name
 
     async def __call__(self, req: Request, res: Response, **params) -> None:
         try:
-            await self._view(req, res, **params)
+            await self.view(req, res, **params)
         except HandlerDoesNotExist as e:
             raise HTTPError(405) from e
 
@@ -148,6 +166,9 @@ class HTTPRouter(HTTPApp, BaseRouter[HTTPRoute]):
 
     Note: routes are stored by `name` instead of `pattern`.
     """
+
+    def _get_key(self, route: HTTPRoute) -> str:
+        return route.name
 
     def add_route(
         self,
@@ -203,7 +224,7 @@ class HTTPRouter(HTTPApp, BaseRouter[HTTPRoute]):
             name = namespace + ":" + name
 
         route = HTTPRoute(pattern=pattern, view=view, name=name)
-        self.routes[name] = route
+        self.add(route)
 
         return route
 
@@ -236,7 +257,7 @@ class WebSocketRoute(BaseRoute):
 
     def __init__(self, pattern: str, view: WebSocketView, **kwargs):
         super().__init__(pattern)
-        self._view = view
+        self.view = view
         self._ws_kwargs = kwargs
 
     async def __call__(
@@ -244,7 +265,7 @@ class WebSocketRoute(BaseRoute):
     ):
         ws = WebSocket(scope, receive, send, **self._ws_kwargs)
         try:
-            await self._view(ws, **params)
+            await self.view(ws, **params)
         except BaseException:
             await ws.ensure_closed(1011)
             raise
@@ -255,6 +276,9 @@ class WebSocketRouter(BaseRouter[WebSocketRoute]):
 
     Extends [BaseRouter](#baserouter).
     """
+
+    def _get_key(self, route: WebSocketRoute) -> str:
+        return route.pattern
 
     def add_route(self, view: WebSocketView, pattern: str, **kwargs):
         """Register a WebSocket route.
@@ -267,7 +291,7 @@ class WebSocketRouter(BaseRouter[WebSocketRoute]):
         route (WebSocketRoute): the registered route.
         """
         route = WebSocketRoute(pattern=pattern, view=view, **kwargs)
-        self.routes[pattern] = route
+        self.add(route)
         return route
 
     async def __call__(self, scope: Scope, receive: Receive, send: Send):
@@ -279,3 +303,115 @@ class WebSocketRouter(BaseRouter[WebSocketRoute]):
             await WebSocketClose(code=403)(receive, send)
         else:
             await match.route(scope, receive, send, **match.params)
+
+
+class RoutingMixin:
+    """Provide HTTP and WebSocket routing to an application class."""
+
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+        self.http_router = HTTPRouter()
+        self.websocket_router = WebSocketRouter()
+
+    def route(self, pattern: str, *, name: str = None, namespace: str = None):
+        """Register a new route by decorating a view.
+
+        # Parameters
+        pattern (str): an URL pattern.
+        methods (list of str):
+            An optional list of HTTP methods.
+            Defaults to `["get", "head"]`.
+            Ignored for class-based views.
+        name (str):
+            An optional name for the route.
+            If a route already exists for this name, it is replaced.
+            Defaults to a snake-cased version of the view's name.
+        namespace (str):
+            An optional namespace for the route. If given, it is prefixed to
+            the name and separated by a colon.
+
+        # See Also
+        - [check_route](#check-route) for the route validation algorithm.
+        """
+        return self.http_router.route(
+            pattern=pattern, name=name, namespace=namespace
+        )
+
+    def websocket_route(
+        self,
+        pattern: str,
+        *,
+        value_type: Optional[str] = None,
+        receive_type: Optional[str] = None,
+        send_type: Optional[str] = None,
+        caught_close_codes: Optional[Tuple[int]] = None,
+    ):
+        """Register a WebSocket route by decorating a view.
+
+        # Parameters
+        pattern (str): an URL pattern.
+
+        # See Also
+        - [WebSocket](./websockets.md#websocket) for a description of keyword
+        arguments.
+        """
+        # NOTE: use named keyword arguments instead of `**kwargs` to improve
+        # their accessibility (e.g. for IDE discovery).
+        return self.websocket_router.route(
+            pattern=pattern,
+            value_type=value_type,
+            receive_type=receive_type,
+            send_type=send_type,
+            caught_close_codes=caught_close_codes,
+        )
+
+    def url_for(self, name: str, **kwargs) -> str:
+        """Build the URL path for a named route.
+
+        # Parameters
+        name (str): the name of the route.
+        kwargs (dict): route parameters.
+
+        # Returns
+        url (str): the URL path for a route.
+
+        # Raises
+        HTTPError(404) : if no route exists for the given `name`.
+        """
+        route = self.http_router.routes.get(name)
+        if route is None:
+            raise HTTPError(404)
+        return route.url(**kwargs)
+
+    def redirect(
+        self,
+        *,
+        name: str = None,
+        url: str = None,
+        permanent: bool = False,
+        **kwargs,
+    ):
+        """Redirect to another HTTP route.
+
+        # Parameters
+        name (str): name of the route to redirect to.
+        url (str):
+            URL of the route to redirect to (required if `name` is omitted).
+        permanent (bool):
+            If `False` (the default), returns a temporary redirection (302).
+            If `True`, returns a permanent redirection (301).
+        kwargs (dict):
+            Route parameters.
+
+        # Raises
+        Redirection:
+            an exception that will be caught to trigger a redirection.
+
+        # See Also
+        - [Redirecting](../guides/http/redirecting.md)
+        """
+        if name is not None:
+            url = self.url_for(name=name, **kwargs)
+        else:
+            assert url is not None, "url is expected if no route name is given"
+        raise Redirection(url=url, permanent=permanent)
